requirements.txt
ultralytics==8.0.180
django==3.2.8
python-dotenv
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
opencv-python==4.7.0.72
djangorestframework
attrs
celery==5.3.0
pytest==7.4.0
pytest-django==4.5.2
freezegun==1.2.2
pandas==1.5.3
numpy==1.22.4
matplotlib==3.7.1
django-cors-headers==4.2.0
pika==1.3.2
scikit-learn==1.3
gym==0.26.2
keras==2.13.1
tensorflow==2.13.0
